Move optype enums to a separate header.
